refactor(ledger): deduplicate route queries
